Use long instead of int for address in DirectByteBuffer constructor
This allows the code to be used on 64-bit platforms.

(cherry picked from commit 9ce6a2533b50ffcd24b8b10bae7f234042768819)

+++ luni/src/main/java/java/nio/DirectByteBuffer.java
@@ -44,7 +44,7 @@ class DirectByteBuffer extends MappedByteBuffer {
   }
 
   // Used by the JNI NewDirectByteBuffer function.
-  DirectByteBuffer(int address, int capacity) {
+  DirectByteBuffer(long address, int capacity) {
     this(MemoryBlock.wrapFromJni(address, capacity), capacity, 0, false, null);
   }
